add ray class
Now we are ready to make a "ray tracer". At the core a ray tracer is
to send rays through pixels and compute what color is seen in the
direction of those rays. Add code to perform something similar.

Trivia: We are currently using a linear interpolation function to
generate this image.